Fix loaded file check before renaming in lowloaderproxy_clean_reloaded_files()

+++ src_template/plnshots.sh
@@ -2277,6 +2277,7 @@ lowloaderproxy_clean_reloaded_files()
         reload_dir="$(echo "$line" | reloadlinehand_proxy_get_dir)"
         reload_file="$(echo "$line" | reloadlinehand_proxy_get_file)"
         load_file="$(sitefpo_make_load_file "$reload_file")"
+        [ -f "$reload_dir/$reload_file" ] || return 1
         rm -f "$reload_dir/$load_file" || return 1
         mv "$reload_dir/$reload_file" "$reload_dir/$load_file" || return 1
     done
